fix conflicts
remove facts

switch to puppet5

update

+++ facts/get_facts.sh
@@ -1,5 +1,4 @@
-#!/bin/bash
-
+#!/bin/sh
 export PATH=/opt/puppetlabs/bin:$PATH
 
 if test -f /usr/bin/apt-get; then
@@ -10,11 +9,13 @@ if test -f /usr/bin/apt-get; then
   operatingsystemmajrelease=$(lsb_release -sr)
   osfamily='Debian'
 elif test -f /usr/bin/dnf; then
-  operatingsystemmajrelease=$(cat /etc/redhat-release | cut -d' ' -f3 )
+  operatingsystemmajrelease=$(awk '{print $3}' /etc/redhat-release)
   osfamily='Fedora'
 elif test -f /usr/bin/yum; then
-  operatingsystemmajrelease=$(cat /etc/redhat-release | cut -d' ' -f4 | cut -c1)
+  operatingsystemmajrelease=$(awk '{print $3}' /etc/redhat-release| cut -c1)
   osfamily='RedHat'
+else
+  osfamily=$(uname)
 fi
 
 case "${osfamily}" in
@@ -25,7 +26,7 @@ case "${osfamily}" in
     dnf install -y "puppet-agent-${puppet_agent_version}.fedoraf${operatingsystemmajrelease}"
     output_file="/vagrant/$(facter --version | cut -c1-3)/$(facter operatingsystem | tr '[:upper:]' '[:lower:]')-$(facter operatingsystemmajrelease)-$(facter hardwaremodel).facts"
     mkdir -p $(dirname ${output_file})
-    facter --show-legacy -p -j | tee ${output_file}
+    [ ! -f ${output_file} ] && facter --show-legacy -p -j | tee ${output_file}
   done
   ;;
 'RedHat')
@@ -35,28 +36,33 @@ case "${osfamily}" in
     yum install -y puppet-agent-${puppet_agent_version}
     output_file="/vagrant/$(facter --version | cut -c1-3)/$(facter operatingsystem | tr '[:upper:]' '[:lower:]')-$(facter operatingsystemmajrelease)-$(facter hardwaremodel).facts"
     mkdir -p $(dirname ${output_file})
-    facter --show-legacy -p -j | tee ${output_file}
+    [ ! -f ${output_file} ] && facter --show-legacy -p -j | tee ${output_file}
   done
   ;;
 
 'Debian')
-  if [[ "xenial" =~ ${lsbdistcodename} ]]; then
-    lsbdistcodename='wily'
-  fi
-  if [[ "serena" =~ ${lsbdistcodename} ]]; then
-    lsbdistcodename='xenial'
-  fi
+  case "${lsbdistcodename}" in
+    *xenial*) lsbdistcodename='wily' ;;
+    *serena*) lsbdistcodename='xenial' ;;
+  esac
   apt-get install -y wget
   wget "https://apt.puppetlabs.com/puppetlabs-release-pc1-${lsbdistcodename}.deb" -O /tmp/puppetlabs-release-pc1.deb
   dpkg --install /tmp/puppetlabs-release-pc1.deb
   apt-get update
-  for puppet_agent_version in 1.2.2 1.4.2 1.5.3; do
+  for puppet_agent_version in 1.10.7 1.2.2 1.4.2 1.5.3 1.10.6;  do
     apt-get -y --force-yes install puppet-agent=${puppet_agent_version}*
     output_file="/vagrant/$(facter --version | cut -c1-3)/$(facter operatingsystem | tr '[:upper:]' '[:lower:]')-$(facter operatingsystemmajrelease)-$(facter hardwaremodel).facts"
     mkdir -p $(dirname ${output_file})
-    facter --show-legacy -p -j | tee ${output_file}
+    [ ! -f ${output_file} ] && facter --show-legacy -p -j | tee ${output_file}
   done
   apt-get install -y make gcc libgmp-dev
+  ;;
+'FreeBSD')
+  pkg update
+  pkg install -y sysutils/puppet5 sysutils/facter
+  output_file="/vagrant/$(facter --version | cut -c1-3)/$(facter operatingsystem | tr '[:upper:]' '[:lower:]')-$(facter operatingsystemmajrelease)-$(facter hardwaremodel).facts"
+  mkdir -p $(dirname ${output_file})
+  [ ! -f ${output_file} ] && facter --show-legacy -p -j | tee ${output_file}
   ;;
 esac
 
@@ -71,7 +77,7 @@ PATH=/opt/puppetlabs/puppet/bin:$PATH
 gem install bundler --no-ri --no-rdoc --no-format-executable
 bundle install --path vendor/bundler
 
-for version in 1.6.0 1.7.0 2.0.0 2.1.0 2.2.0 2.3.0 2.4.0; do
+for version in 1.6.0 1.7.0 2.0.0 2.1.0 2.2.0 2.3.0 2.4.0 2.5.0; do
   FACTER_GEM_VERSION="~> ${version}" bundle update
   case "${operatingsystem}" in
     openbsd)
@@ -81,6 +87,7 @@ for version in 1.6.0 1.7.0 2.0.0 2.1.0 2.2.0 2.3.0 2.4.0; do
       output_file="/vagrant/$(bundle exec facter --version | cut -c1-3)/${operatingsystem}-${operatingsystemmajrelease}-${hardwaremodel}.facts"
       ;;
   esac
+  [ -f ${output_file} ] && continue
   mkdir -p $(dirname $output_file)
   echo $version | grep -q -E '^1\.' &&
     FACTER_GEM_VERSION="~> ${version}" bundle exec facter -j | bundle exec ruby -e 'require "json"; jj JSON.parse gets' | tee $output_file ||
